requirements.txt
GitPython==3.1.43
Jinja2==3.1.3
MarkupSafe==2.1.5
PyHessian==0.1
PySocks==1.7.1
PyYAML==6.0
absl-py==2.1.0
accelerate==0.32.1
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
cmake==3.29.2
contextlib2==21.6.0
contourpy==1.2.1
cycler==0.12.1
datasets==2.21.0
dill==0.3.8
docker-pycreds==0.4.0
evaluate==0.4.2
filelock==3.12.0
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2023.5.0
ftfy==6.1.1
gdown==5.2.0
gitdb==4.0.11
huggingface-hub==0.21.2
idna==3.7
importlib_resources==6.4.0
joblib==1.4.2
kiwisolver==1.4.5
lit==18.1.4
littleutils==0.2.4
matplotlib==3.9.1
ml-collections==0.1.1
mpmath==1.2.1
multidict==6.0.5
multiprocess==0.70.16
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
ogb==1.3.6
open-clip-torch==2.10.1
outdated==0.2.2
packaging==23.1
pandas==2.2.2
pathtools==0.1.2
pillow==10.3.0
pip==22.0.2
platformdirs==4.2.1
promise==2.3
protobuf==3.20.3
psutil==5.9.8
pyarrow-hotfix==0.6
pyarrow==16.1.0
pycparser==2.22
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytorch-ranger==0.1.1
pytorch-triton-rocm==2.1.0
pytz==2024.1
regex==2023.6.3
requests==2.32.3
safetensors==0.3.1
scikit-learn==1.5.1
scipy==1.10.1
sentencepiece==0.1.99
sentry-sdk==2.1.1
setproctitle==1.3.3
setuptools==59.6.0
shortuuid==1.0.13
six==1.16.0
smmap==5.0.1
soupsieve==2.5
sympy==1.12
termcolor==2.3.0
threadpoolctl==3.5.0
timm==0.9.2
tokenizers==0.19.1
torch-optimizer==0.3.0
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tqdm==4.66.4
triton==2.3.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wandb==0.17.0
wcwidth==0.2.6
wheel==0.44.0
wilds==2.0.0
xxhash==3.4.1
yarl==1.9.4
yaspin==3.0.2
zipp==3.19.2